[CORDA-1941]: Started to define API and behaviour for draining shutdown on server side. WIP.
